Move and rename trustchain FFI module, update kotlin version
