+++ lib/app/interop/trustchain/bridge_generated.dart
