+++ test/app/ffi_tests.dart
@@ -1,5 +1,5 @@
 import 'package:credible/app/shared/key_generation.dart';
-import 'package:credible/ffi.dart';
+import 'package:credible/app/interop/trustchain/trustchain.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 void main() {
